auction_admin_bot/requirements.txt
aiogram==3.13.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
apscheduler==3.10.4
python-dotenv==1.0.1
openpyxl==3.1.2
aiohttp==3.10.5
